Move sort routine to library
